soundfront/requirements.txt
autopep8==1.4.4
Click==7.0
Flask==1.0.2
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
pycodestyle==2.5.0
pyodbc==4.0.26
python-dotenv==0.10.1
Werkzeug==0.15.2
